requirements.txt
fastapi==0.63.0
python-dotenv
uvicorn[standard]
psycopg2